lettuce: power_profile: Fudge the radio on current draw
The accounting for radio active time changed heavily on Oreo and
it seems to be giving drastically incorrect but roughly proportional
radio active times. The reported times have increased by an order of
magnitude (and are clearly wrong for my understanding of what radio
active means), so we need to reduce the claimed current draw by an
order of magnitude to balance out the energy accounting.

This change is probably related to the issue:
LineageOS/android_frameworks_base@21f76aa

Change-Id: I2542aa900ad645fb5eb5a7076c71f0dd977332f3